requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
websockets==12.0
aiohttp==3.10.0
python-dotenv==1.0.1
